+++ scripts/helpers.sh
@@ -2,9 +2,6 @@
 default_sessions_dir="$HOME/.tmux/sessions"
 sessions_dir_option="@session-saver-dir"
 
-restore_processes_option="@session-saver-restore-processes"
-default_restore_processes="true"
-
 SUPPORTED_VERSION="1.9"
 
 # helper functions