+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### master
 - restore pane processes
 - user option for disabling pane process restoring
+- enable whitelisting processes that will be restored
 
 ### v0.0.4, 2014-08-26
 - restore pane layout for each window